Added code for Get command along with bug fixes

+++ src/dfs_api/ClientRequestPacket.java
@@ -15,7 +15,7 @@ public class ClientRequestPacket implements Serializable
 
     public String file_name;
     public String arguments[];
-    public int file_size;
+    public double file_size;
 
     public ArrayList<StorageNode> dn_list;
